Docker_Front_Back/backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-cors-headers>=3.10.1
djangorestframework==3.12.4
google-cloud-bigquery